[F] (closed #56) Tooltip in settings `Added:` Number of user messages `Added:` Tooltip that explains number of saved messages `Changed:` ESLint ignore strings `Fixed:` Home page autocomplete
